requirements.dev.txt
# trainer only
torchmetrics
deepspeed
lightning
pytorch_lightning
scikit-learn

# dataset only
torchtext
datasets

# Eval only
pypareto
pyrouge

# Other dependencies
requests
tqdm